Main

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Documentation

data Opcao

Opções usadas nos diferentes menus(Menu Principal, Menu Pausa, Menu Perdeu e Menu Tipo de Jogo)

Constructors

Jogar
Identifica se estamos na opção Jogar

Sair
Identifica se estamos na opção Sair

Menu
Identifica se estamos na opção "Menu Principal"

Guardar
Identifica se estamos na opção "Guardar Jogo"

GuardarS
Identifica se o Jogo foi guardado no Menu Pausa

Novo
Identifica se estamos na opção "Novo Jogo"

Carregar
Identifica se estamos na opção "Carregar Jogo"

data Menu

Diferentes Menus do Jogo

Constructors

MenuP Opcao
Identifica que estamos no Menu Principal e qual a Opcao atual

TipoJ Opcao
Identifica se estamos no Menu Tipo de Jogo e qual a Opcao atual

Pausa Opcao
Identifica se estamos no Menu de Pausa e qual a Opcao atual

Perdeu Opcao
Identifica se estamos no Menu de Perdeu o Jogo e qual a Opcao atual

Instrucoes
Identifica se estamos no Menu de Instruções do Jogo

ModoJogo
Identifica se estamos no Modo de Jogo

type Tempo = Float

Tempo de duração do Jogo

type Pontuacao = (PontuacaoAtual, NivelAtual)

Pontuação do jogo atual

type PontuacaoAtual = Int

Pontuação do jogo atual

type NivelAtual = Int

Nível atual do jogador em relação ao mapa(cada terreno novo corresponde a um nível)

type MelhorPontuacao = Int

Melhor Pontuação obtida em todos os jogos.

type EstadoJogo = (Tempo, Pontuacao, Jogo, Jogada)

Estado de cada Jogo.

type EstadoGloss = (Menu, EstadoJogo, Int, MelhorPontuacao, Imagens)

Estado usado no Gloss.

data Imagens

As diferentes texturas usadas no Jogo separadas por tipo.

Constructors

Texturas

Fields

menu :: [(String, Picture)]

Texturas relativas aos diferentes Menus

jogador :: [(String, Picture)]

Texturas relativas ao Jogador

relva :: [(Obstaculo, Picture)]

Texturas relativas ao terreno Relva

estrada :: [(Obstaculo, Picture)]

Texturas relativas ao terreno Estrada

rio :: [(Obstaculo, Picture)]

Texturas relativas ao terreno Rio

mapaVazio :: Mapa

Função que devolve um mapa com a primeira linha do jogo.

novoJogo :: Int -> EstadoJogo

Função que devolve um EstadoJogo de um novo Jogo.

Usa as funções mapaInicial e mapaVazio como auxíliares para gerar um novo Jogo

leMelhorPontuacao :: IO Int

Função que verifica se existe um ficheiro com a Melhor Pontuação obtida

mapaInicial :: Mapa -> Int -> Int -> Mapa

Função que gera um Mapa inicial aleatorio

estadoInicial :: Imagens -> Int -> Int -> EstadoGloss

Função que gera o estado inicial do Jogo

desenhaEstado :: EstadoGloss -> IO Picture

Função que através de um estado desenha os diferentes menus do Jogo

desenhaJogador :: Jogo -> Jogada -> Imagens -> [Picture]

Função que desenha o Jogador no ecrã conforme a sua posição no Jogo.

desenhaBarraJogo :: Tempo -> Pontuacao -> Coordenadas -> Imagens -> [Picture]

Função que desenha a barra de Tempo e Pontuação do jogo atual

desenhaMapa :: EstadoGloss -> (Int, Int) -> [Picture]

Função que desenha um Mapa no ecrã

desenhaJogoGuardado :: Imagens -> [Picture]

Função que desenha no ecrã se o jogo foi guardado.

verificaSave :: IO (Tempo, Pontuacao, Jogo, Jogada)

Função que verifica se já existe algum jogo guardado para o carregar para o jogo.

window :: Display

Função que cria uma janela de Jogo

reageEvento :: Event -> EstadoGloss -> IO EstadoGloss

Função que devolve um Estado do Gloss conforme uma tecla é pressionada

Dependendo do Estado pode ainda guardar o jogo, carregar jogo guardados e sair do Jogo.

reageTempoGloss :: Float -> EstadoGloss -> IO EstadoGloss

Função que altera o Estado do Gloss conforme o tempo

Passado 5 segundos de Jogo chama a função desliza mapa para dar continuade ao jogo. Verifica ainda se o jogador perdeu o jogo.

fr :: Int

Função de frame rate

valorAleatorio :: Int -> Int

Função para transformar um IO (Int) num Int

main :: IO ()

Função main do Jogo