Tarefa1_2022li1g127

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

tamanhoValido :: Mapa -> Bool

A função tamanhoValido verifica se o Mapa possui um tamanho válido e se a quantidade de Obstaculos corresponde ao comprimento do Mapa.

Exemplos de utilização:

>>> tamanhoValido Mapa 2 [(Relva, [Nenhum, Nenhum]), (Relva, [Nenhum, Nenhum])]
True

>>> tamanho Valido Mapa 1 [(Relva, [Nenhum, Nenhum]), (Relva, [Nenhum, Nenhum])]
False

relvaContiguos :: Int -> Mapa -> Bool

A função relvaContiguos confirma que se quantidade de Terrenos Relva seguidos está dentro dos limites estabelecidos(neste caso, 5).

Exemplos de utilização:

>>> relvaContiguos _ (Mapa 8 [(Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _, [_]),])
True

>>> relvaContiguos _ (Mapa 8 [(Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva _, [_]), (Relva _, [_]), (Rio _, [_]),])
False

estradaContiguos :: Int -> Mapa -> Bool

A função estradaContiguos confirma que a quantidade de Terrenos Estrada seguidos está dentro dos limites estabelecidos(neste caso, 4).

Exemplos de utilização:

>>> estradaContiguos _ (Mapa 8 [(Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_]), (Relva, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _, [_]),])
True

>>> estradaContiguos _ (Mapa 8 [(Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_]), (Estrada _ _, [_]), (Relva _, [_]), (Rio _, [_]),])
False

riosContiguos :: Int -> Mapa -> Bool

A função riosContiguos confirma que a quantidade de Terrenos Rio seguidos está dentro dos limites estabelecidos(neste caso, 4).

Exemplos de utilização:

>>> riosContiguos _ (Mapa 8 [(Rio _, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _, [_]), (Relva, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _, [_]),])
True

>>> riosContiguos _ (Mapa 8 [(Rio _, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _, [_]), (Rio _ _, [_]), (Relva _, [_]), (Rio _, [_]),])
False

terrenosContiguos :: Mapa -> Bool

A função terrenosContiguos utiliza as funções relvaContiguos, estradaContiguos e riosContiguos para verificar que a quantidade de Terrenos seguidos estão dentro dos limites estabelecidos.

Exemplos de utilização:

>>> terrenosContiguos Mapa 6 [(Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Rio _, [_])]
True

>>> terrenosContiguos Mapa 6 [(Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_]), (Relva, [_])]
False

mapaValido :: Mapa -> Bool

A função mapaValido utiliza as funções tamanhoValido, verificaObstaculo e terrenosContiguos para verificar se um Mapa é considerado válido.

Exemplos de utilização:

>>> mapaValido Mapa 2 [(Relva, [Nenhum, Arvore]), (Rio 2, [Tronco, Nenhum])]
True

>>> mapaValido Mapa 2 [(Relva, [Carro, Tronco]), (Rio 3, [Nenhum, Nenhum, Arvore]), (Rio 5, [Nenhum, Nenhum])]
False

riosDirecao :: Int -> Mapa -> Bool

A função riosDirecao verifica que se dois Rio seguidos têm direções opostas.

Exemplos de utilização:

>>> riosDirecao _ (Mapa 2 [(Rio 2, [Nenhum, Tronco]), (Rio (-3), [Tronco, Nenhum])])
True

>>> riosDirecao _ (Mapa 2 [(Rio (-2), [Nenhum, Tronco]), (Rio (-3), [Tronco, Nenhum])])
False

compTroncos :: Int -> [Obstaculo] -> Bool

A função compTroncos verifica se comprimento dos Obstaculo Tronco estão dentro do limite estabelecido (neste caso, 5).

Exemplos de utilização:

>>> compTroncos _ (Mapa 6 [(Rio _, [Tronco, Tronco, Tronco, Tronco, Tronco, Nenhum]), ...])
True

>>> compTroncos _ (Mapa 7 [(Rio _, [Tronco, Tronco, Tronco, Tronco, Tronco, Tronco, Nenhum]), ...])
False

compCarros :: Int -> [Obstaculo] -> Bool

A função compTroncos verifica se comprimento dos Obstaculo Carro estão dentro do limite estabelecido (neste caso, 3).

Exemplos de utilização:

>>> compCarros _ (Mapa 6 [(Estrada _, [Carro, Carro, Carro, Nenhum, Carro, Nenhum]), ...])
True

>>> compCarros _ (Mapa 7 [(Estrada _, [Carro, Carro, Carro, Carro, Nenhum, Carro, Nenhum]), ...])
False

verificaNenhum :: [Obstaculo] -> Bool

A função verificaNenhum verifica a existencia de pelo menos um Obstaculo Nenhum em cada Terreno

Exemplos de utilização:

>>> verificaNenhum [Arvore, Nenhum, Arvore]
True

>>> verificaNenhum [Arvore, Arvore]
False

verificaObstaculo :: Mapa -> Bool

A função verificaObstaculo utiliza as funções obstaculosRelva, obstaculosEstrada, obstaculosRio, verificaNenhum, compCarros, riosDirecao e compTroncos para verificar se os Obstaculo são válidos, i.e. sejam do comprimento válido, estejam no terreno válido, tenham pelo menos um Nenhum, entre outros...

Exemplos de utilização:

>>> verificaObstaculo (Mapa 2 [(Relva, [Arvore, Nenhum]), (Rio 2, [Tronco, Nenhum])])
True

>>> verificaObstaculo (Mapa 2 [(Relva, [Arvore, Arvore]), (Rio 2, [Tronco, Carro])])
False

obstaculosRelva :: [Obstaculo] -> Bool

A função obstaculosRelva verifica se os Obstaculo atuais do Terreno Relva são obstaculos próprios ao Terreno.

Exemplos de utilização:

>>> obstaculosRelva [Arvore, Nenhum, Arvore]
True

>>> obstaculosRelva [Arvore, Nenhum, Carro]
False

obstaculosEstrada :: [Obstaculo] -> Bool

A função obstaculosEstrada verifica se os Obstaculo atuais do Terreno Estrada são obstaculos próprios ao Terreno.

Exemplos de utilização:

>>> obstaculosEstrada [Carro, Nenhum, Carro]
True

>>> obstaculosEstrada [Arvore, Nenhum, Carro]
False

obstaculosRio :: [Obstaculo] -> Bool

A função obstaculosRio verifica se os Obstaculo atuais do Terreno Rio são obstaculos próprios ao Terreno

Exemplos de utilização:

>>> obstaculosRio [Tronco, Nenhum, Tronco]
True

>>> obstaculosRio [Arvore, Nenhum, Tronco]
False