Tarefa2_2022li1g127

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

A função estendeMapa cria um novo Terreno aleatorio no mapa utilizando as funções escolheTerreno e escolheObstaculo.

Exemplos de utilização:

>>> estendeMapa (Mapa 2 [(Relva, [Arvore, Nenhum]), (Rio _, [Nenhum, Tronco])]) _
[(Estrada _, [Carro, Nenhum]), (Rio _, [Nenhum, Tronco])]

>>> estendeMapa (Mapa 2 [(Rio _, [Tronco, Nenhum]), (Rio _, [Nenhum, Tronco])]) _
[(Rio _, [Tronco, Nenhum]), (Rio _, [Nenhum, Tronco])]

alteraVelocidade :: Terreno -> Int -> Int -> Int -> Terreno

A função alteraVelocidade recebe um Terreno e três inteiros e altera o valor da velocidade dos terrenos usando a divisão dos valores inteiros recebidos.

Exemplos de utilização:

>>> alteraVelocidade (Rio 0) 35 5
Rio 7

>>> alteraVelocidade (Relva) 25 8
Relva

escolheTerreno :: Mapa -> Int -> Terreno

A função escolheTerreno utiliza a ajuda das funções proximosTerrenosValidos, funcaoValorAleatorio e alteraVelocidade para escolher um Terreno de uma lista de próximos Terrenos válidos.

Exemplos de utilização:

>>> escolheTerreno (Mapa 2 [(Relva, [Nenhum, Arvore]), (Relva, [Nenhum, Arvore])]) 22
[(Relva, [Arvore, Nenhum]), (Relva, [Nenhum, Arvore])]

>>> escolheTerreno (Mapa 2 [(Relva, [Nenhum, Arvore]), (Relva, [Nenhum, Arvore])]) 82
[(Rio _, [Tronco, Nenhum]), (Relva, [Nenhum, Arvore])]

escolheObstaculo :: Int -> Int -> Int -> Terreno -> [Obstaculo] -> [Obstaculo]

A função escolheObstaculo utiliza a ajuda das funções proximosObstaculosValidos, funcaoValorAleatorio para para escolher um Obstaculo de uma lista de próximos Obstaculo válidos.

Exemplos de utilização:

>>> escolheObstaculo (2)(1)(_)(Relva)[]
[Nenhum, Arvore]

>>> escolheObstaculo (2)(1)(_)(Rio _)[]
[Nenhum, Tronco]

funcaoValorAleatorio :: Int -> Int -> Int -> Int

A função funcaoValorAleatorio utiliza a função mod para gerar um numero pseudo-aleatorio atráves do tamanho da lista.

Exemplos de utilização:

>>> funcaoValorAleatorio 8 2 2
0

>>> funcaoValorAleatorio 5 2 3
1

proximosTerrenosValidos :: Mapa -> [Terreno]

A função proximosTerrenosValidos utiliza a ajuda das funções relvaContiguos, estradaContiguos e riosContiguos para verificar qual Terreno pode ser gerado dentro das restrições.

Exemplos de utilização:

>>> proximosTerrenosValidos (Mapa 4 [(Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_]), (Estrada _, [_])])
[Rio _, Relva]

>>> proximosTerrenosValidos (Mapa 4 [(Rio _, [_]), (Rio _, [_]), (Rio _, [_]), (Estrada _, [_])])
[Rio _, Estrada _, Relva]

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidos utiliza a ajuda das funções verificaNenhum, compTroncos e compCarros para verificar qual Obstaculo pode ser gerado dentro das restrições.

Exemplos de utilização:

>>> proximosObstaculosValidos 4 (Relva, [])
[Arvore, Nenhum]

>>> proximosObstaculosValidos 2 (Rio _, [Tronco])
[Nenhum]