Tarefa3_2022li1g127

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo utiliza as posicaoJogador e animaMapa para "animar" um Mapa.

Exemplos de utilização:

>>> animaJogo(Jogo(Jogador(1,1)) (Mapa 3 [(Rio 1, [Nenhum, Tronco, Tronco]), (Relva, [Arvore, Nenhum, Nenhum]), (Estrada (-1), [Carro, Nenhum, Nenhum])])) (Move Direita)
Jogo(Jogador(2,1)) (Mapa 3 [(Rio 1, [Nenhum, Tronco, Tronco]), (Relva, [Arvore, Nenhum, Nenhum]), (Estrada (-1), [Carro, Nenhum, Nenhum])])

>>> animaJogo(Jogo(Jogador(1,1)) (Mapa 3 [(Rio 1, [Nenhum, Tronco, Tronco]), (Relva, [Arvore, Nenhum, Nenhum]), (Estrada (-1), [Carro, Nenhum, Nenhum])])) (Move Cima)
Jogo(Jogador(1,0)) (Mapa 3 [(Rio 1, [Nenhum, Tronco, Tronco]), (Relva, [Arvore, Nenhum, Nenhum]), (Estrada (-1), [Carro, Nenhum, Nenhum])])

moveJogador :: Jogador -> Jogada -> Jogador

A função moveJogador move o Jogador uma unidade na direção pretendida.

Exemplos de utilização:

>>> moveJogador (Jogador (1,1)) (Move Cima)
Jogador (1,0)

>>> moveJogador (Jogador (2,2)) (Move Esquerda)
Jogador (1,2)

posicaoJogador :: Mapa -> Jogador -> Jogada -> Jogador

A função posicaoJogador move o Jogador no Mapa tendo em conta as restrições de movimento, utilizando as funções auxiliares posicaoMapa, moveJogadorTronco, moveJogador, obstaculoDirecao e verificaMovValido.

Exemplos de utilização:

>>> posicaoJogador (Mapa 2 [(Relva, [Arvore, Nenhum]), (Relva, [Nenhum, Arvore])] (Jogador (1,0)) (Parado))
Jogador (1,0)

>>> posicaoJogador (Mapa 2 [(Rio (-1), [Nenhum, Tronco]), (Relva, [Nenhum, Arvore])] (Jogador (1,0)) (Parado))
Jogador (0,0)

moveJogadorTronco :: Mapa -> Jogador -> Jogada -> Jogador

A função moveJogadorTronco move o Jogador na velocidade e na direção do Obstaculo Tronco, utilizando as funções auxiliares verificaTerrenoCoord, velocidadeRio, posicaoMapa, obstaculoDirecao e verificaMovValido.

Exemplos de utilização:

>>> moveJogadorTronco (Mapa 2 [(Rio (-1), [Nenhum, Tronco]), (Relva, [Nenhum, Arvore])] (Jogador (1,0)) (Parado))
Jogador (0,0)

>>> posicaoJogador (Mapa 2 [(Rio (1), [Nenhum, Tronco]), (Relva, [Nenhum, Arvore])] (Jogador (0,0)) (Parado))
Jogador (1,0)

obstaculoDirecao :: Mapa -> Jogador -> Jogada -> Obstaculo

A função obstaculoDirecao devolve os Obstaculos presentes na posição do Movimento do Jogador, utiliza a função auxiliar posicaoMapa.

Exemplos de utilização:

>>> obstaculoDirecao (Mapa 2 [(Relva, [Nenhum, Arvore]),(Relva, [Nenhum, Arvore])] (Jogador (0,0)) (Parado))
Nenhum

>>> obstaculoDirecao (Mapa 2 [(Relva, [Nenhum, Arvore]),(Relva, [Nenhum, Arvore])] (Jogador (0,0)) (Move Direita))
Arvore

velocidadeRio :: (Terreno, [Obstaculo]) -> Int

A função velocidadeRio devolve a velocidade de um Terreno Rio.

Exemplos de utilização:

>>> velocidadeRio (Rio 3, _)
3

>>> velocidadeRio (Rio (-2), _)
-2

animaMapa :: Jogador -> [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

A função animaMapa anima os Obstaculo de um Mapa conforme a sua velocidade de direção.

Exemplos de utilização:

>>> animaMapa _ []
[]

>>> animaMapa _ [(Rio 1, [Tronco, Tronco, Nenhum])]
[Tronco, Nenhum, Tronco]

>>> animaMapa (Jogador (1,0)) [(Estrada 2, [Carro, Nenhum, Nenhum])]
[Nenhum, Carro, Nenhum]

animaEstrada :: Int -> Int -> [Obstaculo] -> [Obstaculo]

A função animaEstrada move os obstaculos do Terreno Estrada tendo em conta a posição do Jogador.

Exemplos de utilização:

>>> animaEstrada _ _ []
[]

>>> animaEstrada 1 2 [Carro, Nenhum, Nenhum]
[Nenhum, Carro, Nenhum]

>>> animaEstrada 0 1 [Nenhum, Carro, Nenhum]
[Nenhum, Nenhum, Carro]

animaObstaculo :: [Obstaculo] -> Int -> [Obstaculo]

A função animaObstaculo move os Obstaculo conforme a velocidade do Terreno.

Exemplos de utilização:

>>> animaObstaculo [Carro, Carro, Nenhum] 0
[Carro, Carro, Nenhum]

>>> animaObstaculo [Tronco, Tronco, Nenhum] 1
[Tronco, Nenhum, Tronco]

verificaMovValido :: Mapa -> Jogador -> Jogada -> Bool

A função verificaMovValido verifica se um dado movimento do jogador é valido, utiliza as funções auxiliares verificaMovDentroMapa, e verificaObsPorMov.

Exemplos de utilização:

>>> verificaMovValido (Mapa 2 [(Relva, [Nenhum, Arvore]),(Relva, [Nenhum, Arvore])] (Jogador (0,0)) (Move Baixo))
True

>>> verificaMovValido (Mapa 2 [(Relva, [Nenhum, Arvore]),(Relva, [Nenhum, Arvore])] (Jogador (0,0)) (Move Direita))
False

verificaObsPorMov :: Mapa -> Jogador -> Jogada -> Bool

A função verificaObsPorMov verifica se o Jogador não se movimenta para um Obstaculo Arvore, utilizando a função auxiliar posicaoMapa.

Exemplos de utilização:

>>> verificaObsPorMov _ _ Parado
True

>>> verificaObsPorMov (Mapa 2 [(Relva, [Nenhum, Arvore]), (Relva, [Nenhum, Arvore])] (Jogador (0,0))) (Move Direita)
False

posicaoMapa :: Mapa -> Coordenadas -> Obstaculo

A função posicaoMapa devolve um Obstaculo conforme uma coordenada do mapa, utiliza as funções auxiliares verificaObstaculoCoord, e verificaTerrenoCoord.

Exemplos de utilização:

>>> posicaoMapa (Mapa 2 [(Relva, [Arvore, Nenhum]), (Relva, [Nenhum, Arvore])] (0,0))
Arvore

>>> posicaoMapa (Mapa 2 [(Relva, [Arvore, Nenhum]), (Rio _, [Nenhum, Tronco])] (1,1))
Tronco

verificaTerrenoCoord :: Mapa -> Int -> (Terreno, [Obstaculo])

A função verificaTerrenoCoord devolve um tuplo (Terreno, [Obstaculo]) conforme uma Coordenadas do Mapa.

Exemplos de utilização:

>>> verificaTerrenoCoord (Mapa 2 [(Relva, [Nenhum, Arvore]), (Relva, [Nenhum, Arvore])]) 0
(Relva, [Nenhum, Arvore])

>>> verificaTerrenoCoord (Mapa 2 [(Relva, [Nenhum, Arvore]), (Relva, [Nenhum, Arvore])]) 1
(Relva, [Nenhum, Arvore])

verificaObstaculoCoord :: (Terreno, [Obstaculo]) -> Int -> Obstaculo

A função verificaObstaculoCoord dado um tuplo (Terreno, [Obstaculo]) devolve um Obstaculo conforme as Coordenadas do Mapa.

Exemplos de utilização:

>>> verificaObstaculoCoord (Relva, [Nenhum, Arvore]) 0
Nenhum

>>> verificaObstaculoCoord (Relva, [Nenhum, Arvore]) 1
Arvore

verificaMovDentroMapa :: Int -> Jogador -> Jogada -> Bool

A função verificaMovDentroMapa verifica se o Jogador se movimenta dentro dos limites de um Mapa.

Exemplos de utilização:

>>> verificaMovDentroMapa 2 (Jogador (2,2)) (Move Direita)
False

>>> verificaMovDentroMapa 2 (Jogador (1,1)) Parado
True